% Generated by roxygen2 (4.0.1): do not edit by hand
\name{chooseClassifier}
\alias{chooseClassifier}
\title{Choose The Best Classifier}
\usage{
chooseClassifier(formula, train, test)
}
\arguments{
  \item{formula}{A \code{formula} for classifiers to
  compute.}

  \item{train}{A \code{data.frame} denoting the training
  set.}

  \item{test}{A \code{data.frame} denoting the test set.}
}
\description{
For an object of class \code{formula} and a data set the
\code{chooseClassifier} computes a \link{ggplot} showing
the Area Under the Curve (result of
\link[ROCR]{performance}) for classification algorithms
like \link{naiveBayes}, \link{lda}, logistic regression
\link{glm} with \code{logit} or \code{probit} link
function. Any classifier can be added if needed. Contact:
\href{https://github.com/sandersone/twd1/issues}{https://github.com/sandersone/twd1/issues}.
}
\note{
For naive Bayes a parameter \code{laplace} is set to 0.2.
}
\examples{
library(foreign)
se <- read.arff("http://archive.ics.uci.edu/ml/machine-learning-databases/00266/seismic-bumps.arff")
index <- 1:(2*nrow(se)/3)
se <- se[,-c(14:16)]
se_wyb <- se[,-c(9,15)]
train <- se_wyb[ index, ]
test <- se_wyb[ -index, ]
chooseClassifier( class~., train, test )
}

