% Generated by roxygen2 (4.0.1): do not edit by hand
\name{chooseClassifier}
\alias{chooseClassifier}
\title{Choose The Best Classifier}
\usage{
chooseClassifier(formula, train, test)
}
\arguments{
  \item{formula}{\code{formula}}

  \item{train}{\code{data.frame}}

  \item{train}{\code{data.frame}}
}
\description{
For an object of class \code{formula} and a data set the
\code{chooseClassifier} computes a \link{ggplot} showing
the Area Under Curve (result of \link[performance]{ROCR})
for classification algorithms like \link{naiveBayes},
\link{lda}, logistic regression \link{glm} with
\code{logit} link function. Any classifier can be added if
needed. Contact:
\href{https://github.com/sandersone/twd1/issues}{https://github.com/sandersone/twd1/issues}.
}
\details{
TO DO
}
\note{
For k-nearest neighbourhood alghoritm, the model with the
highest Area Under Curve is shown for \code{k} from 2 to 6.
So far it is commented because I have problems with that
function.

For naive Bayes a parameter \code{laplace} is set to 0.2.
}
\examples{
library(foreign)
se <- read.arff("http://archive.ics.uci.edu/ml/machine-learning-databases/00266/seismic-bumps.arff")
index <- 1:(2*nrow(se)/3)
se <- se[,-c(14:16)]
se_wyb <- se[,-c(9,15)]
train <- se_wyb[ index, ]
test <- se_wyb[ -index, ]
chooseClassifier( class~., train, test )
}

